fix: cast messagesize as uint (credit: kaen / #1592) (#1686)

+++ com.unity.netcode.gameobjects/Runtime/Messaging/MessagingSystem.cs
@@ -377,7 +377,7 @@ namespace Unity.Netcode
 
             var header = new MessageHeader
             {
-                MessageSize = (ushort)tmpSerializer.Length,
+                MessageSize = (uint)tmpSerializer.Length,
                 MessageType = m_MessageTypes[typeof(TMessageType)],
             };
             BytePacker.WriteValueBitPacked(headerSerializer, header.MessageType);
